revert

+++ docker/ci/test-v2/build.sh
@@ -10,6 +10,8 @@ proxy_args+=" --build-arg HTTPS_PROXY=${HTTPS_PROXY}"
 proxy_args+=" --build-arg http_proxy=${http_proxy}"
 proxy_args+=" --build-arg https_proxy=${https_proxy}"
 
-img_tag="oneflow-test-v2:$USER" # update me if any of related files are changed
-docker build --rm $proxy_args \
+img_tag="oneflow-test-v2:0.1" # update me if any of related files are changed
+if [[ "$(docker images -q ${img_tag} 2> /dev/null)" == "" ]]; then
+  docker build --rm $proxy_args \
     -t $img_tag .
+fi
